fix: moving role_builder into role directory
