+++ backend/internal/handlers/role/role_builder.go
@@ -1,12 +1,14 @@
-package models
+package role
+
+import "inside-athletics/internal/models"
 
 type PermissionSpec struct {
-	Action   PermissionAction
+	Action   models.PermissionAction
 	Resource string
 }
 
 type RoleSpec struct {
-	Name        RoleName
+	Name        models.RoleName
 	Permissions []PermissionSpec
 }
 
@@ -14,13 +16,13 @@ type RoleBuilder struct {
 	spec RoleSpec
 }
 
-func NewRoleBuilder(name RoleName) *RoleBuilder {
+func NewRoleBuilder(name models.RoleName) *RoleBuilder {
 	return &RoleBuilder{
 		spec: RoleSpec{Name: name},
 	}
 }
 
-func (b *RoleBuilder) WithPermission(action PermissionAction, resource string) *RoleBuilder {
+func (b *RoleBuilder) WithPermission(action models.PermissionAction, resource string) *RoleBuilder {
 	b.spec.Permissions = append(b.spec.Permissions, PermissionSpec{
 		Action:   action,
 		Resource: resource,
